+++ src/command_modules/azure-cli-vm/azure/cli/command_modules/vm/_client_factory.py
@@ -78,3 +78,7 @@ def cf_images(_):
 
 def cf_run_commands(_):
     return _compute_client_factory().virtual_machine_run_commands
+
+
+def cf_rolling_upgrade_commands(_):
+    return _compute_client_factory().virtual_machine_scale_set_rolling_upgrades