+++ src/command_modules/azure-cli-vm/azure/cli/command_modules/vm/_template_builder.py
@@ -684,7 +684,7 @@ def build_vmss_storage_account_pool_resource(loop_name, location, tags, storage_
     return storage_resource
 
 
-# pylint: disable=too-many-locals
+# pylint: disable=too-many-locals, too-many-branches, too-many-statements
 def build_vmss_resource(name, naming_prefix, location, tags, overprovision, upgrade_policy_mode,
                         vm_sku, instance_count, ip_config_name, nic_name, subnet_id,
                         public_ip_per_vm, vm_domain_name, dns_servers, nsg,
@@ -694,7 +694,7 @@ def build_vmss_resource(name, naming_prefix, location, tags, overprovision, upgr
                         image_data_disks, os_type,
                         image=None, admin_password=None, ssh_key_value=None, ssh_key_path=None,
                         os_publisher=None, os_offer=None, os_sku=None, os_version=None,
-                        backend_address_pool_id=None, inbound_nat_pool_id=None,
+                        backend_address_pool_id=None, inbound_nat_pool_id=None, health_probe=None,
                         single_placement_group=None, custom_data=None, secrets=None, license_type=None, zones=None):
 
     # Build IP configuration
@@ -829,6 +829,9 @@ def build_vmss_resource(name, naming_prefix, location, tags, overprovision, upgr
 
     if license_type:
         vmss_properties['virtualMachineProfile']['licenseType'] = license_type
+
+    if health_probe and supported_api_version(ResourceType.MGMT_COMPUTE, min_api='2017-03-30'):
+        vmss_properties['virtualMachineProfile']['networkProfile']['healthProbe'] = {'id': health_probe}
 
     if supported_api_version(ResourceType.MGMT_COMPUTE, min_api='2016-04-30-preview'):
         vmss_properties['singlePlacementGroup'] = single_placement_group