+++ src/command_modules/azure-cli-vm/HISTORY.rst
@@ -4,6 +4,7 @@ Release History
 ===============
 (unreleased)
 +++++++++++++++++++
+* `vm encryption enable`: allow updating encryption settings by rerunning the command
 * `vm create`: expose --os-disk-size-gb
 * `vmss create`: expose --license-type for windows os