+++ src/command_modules/azure-cli-vm/HISTORY.rst
@@ -2,7 +2,7 @@
 
 Release History
 ===============
-(unreleased)
+2.0.16 (2017-10-09)
 +++++++++++++++++++
 * `vm show`: fix a bug when using '-d' crashes on missing private ip addresses
 * `vmss create`: (PREVIEW) support rolling upgrade